refactor: Reorganize Chainlit app into modular frontend architecture
- Created frontend/ directory with 8 modular components:
  * config.py: Configuration, constants, and model filtering
  * cache.py: Dataframe caching with invalidation
  * data_utils.py: Data loading, schema, and profiling utilities
  * knowledge_utils.py: Knowledge store integration
  * commands.py: Slash command handling
  * actions.py: Button action callbacks
  * steps.py: Analysis pipeline steps
  * handlers.py: Main event handlers (on_chat_start, on_message, etc.)

- Created app.py as new entry point (replaces chainlit_app.py)
- Backed up original chainlit_app.py to chainlit_app.py.bak
- Updated all documentation:
  * .replit configuration
  * README.md file structure
  * docs/design.md
  * replit.md

Benefits:
- Improved code organization and maintainability
- Better separation of concerns
- Easier testing and debugging
- Modular components can be reused
- Clearer import structure

All functionality preserved - no breaking changes to the user interface.
