+++ .replit
@@ -25,7 +25,7 @@ author = "agent"
 
 [[workflows.workflow.tasks]]
 task = "shell.exec"
-args = "chainlit run chainlit_app.py --host 0.0.0.0 --port 5000"
+args = "chainlit run app.py --host 0.0.0.0 --port 5000"
 waitForPort = 5000
 
 [workflows.workflow.metadata]
